requirements.txt
Flask==2.0.2
numpy==1.21.4
psycopg2-binary==2.9.2
pyTelegramBotAPI==4.2.0
python-snap7==1.1
requests==2.26.0
